+++ src/FScript.Language/Lexer.fs
@@ -30,6 +30,7 @@ module Lexer =
         | "elif" -> Some Elif
         | "raise" -> Some Raise
         | "for" -> Some For
+        | "import" -> Some Import
         | "include" -> Some Include
         | "module" -> Some Module
         // 'in' is intentionally not a keyword in FScript layout-style let expressions.